+++ finality-provider/service/app.go
@@ -549,13 +549,6 @@ func (app *FinalityProviderApp) setFinalityProviderSlashed(fpi *FinalityProvider
 	}
 }
 
-func (app *FinalityProviderApp) setFinalityProviderJailed(fpi *FinalityProviderInstance) {
-	fpi.MustSetStatus(proto.FinalityProviderStatus_JAILED)
-	if err := app.removeFinalityProviderInstance(); err != nil {
-		panic(fmt.Errorf("failed to terminate a jailed finality-provider %s: %w", fpi.GetBtcPkHex(), err))
-	}
-}
-
 // NOTE: this is not safe in production, so only used for testing purpose
 func (app *FinalityProviderApp) getFpPrivKey(fpPk []byte) (*btcec.PrivateKey, error) {
 	record, err := app.eotsManager.KeyRecord(fpPk, "")