+++ CHANGELOG.md
@@ -40,6 +40,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Improvements
 
 * [#253](https://github.com/babylonlabs-io/finality-provider/issues/253) Refactor to start from the last finalized height
+* [#260](https://github.com/babylonlabs-io/finality-provider/pull/260) Allow running of jailed fp
 
 ## v0.14.2